% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towns.R
\docType{data}
\name{towns}
\alias{towns}
\title{Taiwanese township populations}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 368 rows and 2 columns.
}
\source{
\url{https://data.gov.tw/dataset/8410}
}
\usage{
towns
}
\description{
A dataset of the names and populations of townships/cities and districts in Taiwan (2020).
}
\keyword{datasets}
